+++ source/rr/src/sounds.cpp
@@ -915,9 +915,9 @@ void S_Update(void)
     } while (++sndnum <= highest);
 }
 
-void S_Callback(uint32_t num)
+void S_Callback(intptr_t num)
 {
-    if ((int32_t)num == MUSIC_ID)
+    if (num == MUSIC_ID)
         return;
 
     dq[dnum & (DQSIZE - 1)] = num;
